requirements.txt
matplotlib==3.7.2
numpy==1.24.3
opencv-python==4.7.0.72
pandas==2.0.1
Pillow==9.5.0
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.10.1
seaborn==0.12.2
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.65.0
ultralytics==8.0.131